디버깅용 로그 제거

+++ src/main/java/com/pocketco/domain/user/dto/MainScreenResponse.java
@@ -10,5 +10,5 @@ public record MainScreenResponse(
         Long languageId,
         String languageName,
         List<MainScreenCalenderDTO> totalSolvedDetails,
-        Long thisMonthSolvedCount
+        Integer thisMonthSolvedCount
 ) { }
